h3: limit size of stream state buffer
A peer can influence the size of the state buffer (e.g. with the
payload size of a GREASE frame), so we need to limit the maximum
size to avoid DoS.

In the future we might want to have a dynamic limit set by the
application, but for now a static one will do.