limit stream send buffer size
In order to avoid buffering an infinite amount of data in the stream's
send buffer, we now only allow streams to buffer outgoing data up to the
amount that the peer allows us to send.

This also adds a new Connection method, stream_capacity() which can be
used by applications to know how much data they can write in advance.

This is required by the HTTP/3 code that needs to know the size of body
it can write in advance in order to set a proper size in the DATA frame
header. Without stream_capacity(), send_body() could potentially write
an incomplete DATA frame in the stream send buffer which would be
complicated to recover from.