stream: simplify the Readable iterator
All users of Readable also do a collect() (otherwise they wouldn't be
able to reference the Connection while iterating), so we might as well
do it directly in the API, so we can simplify some code.

This also restores how the Readable iterator is accessed from FFI, to
before commit 67ee5f3b5dff7745abae130e8215ec7e977a5d10. The side effect
of that commit was that a user couldn't ignore a readable stream, and
every call to quiche_readable_next() would keep returning the same
stream. The new API makes it more consistent with Rust.