h3: streamline poll()
This changes the poll() function such that frames received on a stream
are handled immediately, rather than being queued for later processing.

This simplies the general architecture of the HTTP/3 implementation, and
removes a slight potential for abuse, since the stream's frame queue
didn't have a limit. In addition this removes the need to loop through
all streams when poll() is called, to handle previously queued frames.

Benchmarks show some improvement, though this also enables further
optimizations in the future:

  http3/128000            time:   [3.1744 ms 3.3412 ms 3.4806 ms]
                          change: [-19.469% -10.427% -0.5146%] (p = 0.04 < 0.05)
                          Change within noise threshold.
  http3/256000            time:   [11.336 ms 11.527 ms 11.689 ms]
                          change: [-10.013% -7.8772% -5.7020%] (p = 0.00 < 0.05)
                          Performance has improved.
  http3/512000            time:   [25.405 ms 25.766 ms 26.072 ms]
                          change: [-17.732% -16.439% -14.968%] (p = 0.00 < 0.05)
                          Performance has improved.
  http3/1000000           time:   [62.185 ms 63.543 ms 64.738 ms]
                          change: [-8.1289% -6.4240% -4.4967%] (p = 0.00 < 0.05)
                          Performance has improved.

As can be observed by the tests, there is a slight change in behavior,
as events might be returned even if there is a later error (though we
couldn't really depend on the previous behavior, as it depended on when
streams were received).

In addition, while before we would return events in order of stream ID
(as streams were processed in order due to the use of a BTreeMap), the
events order now depends on the order of streams returned by the
readable stream iterator (which is kinda random).

This required some refactoring of HTTP/3 internals:

 * Before this change, when trying to parse unknown frames,
   Frame::from_bytes would return Error::Done, but it now returns a
   special Unknown frame type. Incidentally, this fixes a problem we
   had before, where try_consume_frame() would ignore malformed frames.

 * Finished streams are queued in a separate queue, so that we can
   properly report Finished events.

 * Control streams are now processed separately, before any other
   readable stream.

 * Frame processing is now performed in a separate method.

This also incorporates https://github.com/cloudflare/quiche/pull/85.