h3: prefer config mutation
Previously, the HTTP/3 Config object was constructed with
an explicit set of method parameters and could not be mutated. This
differed from the transport Config object, which is constructed
with minimal parameters and mutated by the application to its needs.

This change causes the HTTP/3 Config object to be constructed empty
and offers metheds to mutate it. This unfortanately is a break change
to the API, but it reduces the chance of future breaking changes.