discard 1-RTT packets received before handshake completion
Even if we only decrypt the packets but not process the frames they
carry, there is still some risk for side effects (see e.g.
cfc21ff96c8654929224aebb54a1673fb55ce7bf).

In the future we may want to buffer received packets so we can later
process them as an optimization.