nginx: delay closure of blocked HTTP/3 streams with no body
After an HTTP/3 request is complete, the respective QUIC stream is
closed. When the stream is blocked, and the headers or the body weren't
completely sent, the stream should not be closed, until all the data is
sent.

This currently works for streams of responses that have a body, because
the data-sending API (send_chain()) will return NGX_AGAIN, and NGINX
will know that it needs to repeat the operation. However, when there is
no body, NGINX will happily close the stream before the headers have
been sent.

This fixes that by delaying the stream closure even in case of no-body
responses.