qlog: update to v0.2.0
Previously, qlog API was designed to operate
in a buffered mode. This required users
to create a trace and hold it in memory,
append events to it, and at the end
serialize the whole thing.

This change adds a streaming mode API that
serializes qlog events immediately to a
`Write` trait.

A few quality-of-life changes have also
been made.