apps: make 'qlog' feature really optional
Both the apps crate and quiche itself optionally depend on the qlog
crate, however the apps crate enables quiche's qlog feature
unconditionally. This means that even if the qlog feature of the apps
crate is disabled, the qlog crate will still be built.

However the apps crate doesn't need to depend on the qlog crate at all,
as it doesn't use it directly. This change removes the dependency, and
turns the qlog feature into an explicit feature that only enables
quiche's qlog feature when enabled.