immediately mark connection as closed when receiving an invalid first packet
When an invalid first packet is received the connection is correctly
marked as closing, however due to the anti-amplification mechanism it is
prevented from actually sending anything.

Since the draining timeout is armed only when `CONNECTION_CLOSE` is
actually sent, and the idle timeout is armed when a packet is processed
successfully, the connection ends up never being closed.