allow sending 0.5 RTT data when 0-RTT is used
When accepting early data, the server needs to be able to send data in
the 0.5 RTT before the handshake is completed, otherwise there would be
no difference from normal 1-RTT data handling.

Note that this doesn't handle the HTTP/3 case right now. That is, hTTP/3
control streams are still only opened after the handshake is completed
when not in 0-RTT.