lib: allow incoming Initial packets lost/reordered
Let's think the following scenario:

0. Client sends Initial pn=0
1. Server sends Initial pn=0 -> Lost
2. Server sends Handshake pn=0

Currently when Initial#0 is lost, Client received Handshake#0 and
close the connection immediately because now it gets `Error::CryptoFail`
because Handshake#0 cannot be decrypted.

This can happen when 1 is lost, or reordered so 2 is delivered before 1.
Server doesn't know the connection is closed in the Client,
so keep trying to retransmit and timed out eventually.

The fix is to always ignore invalid packets on the client-side, as the
client will already have the idle timer armed on account of having sent
the Initial packet.

Server behavior is not changed(close the connection when invalid packet
is received if not other packet was previously successfully processed).

Test Update:
- invalid_initial() is renamed to invalid_initial_server().
- invalid_initial_client() is added to check the client behavior.