tls: configure early data context
This is required to fix accepting early data, that was broken by the
last BoringSSL update (but only for some implementations for some
reasons, so this wasn't caught at the time). BoringSSL also needs to be
updated as well, because the logic from the previous update was wrong.

The "early data context" is used by BoringSSL to decide whether early
data should be accepted or not. In theory this should include the
transport parameters and HTTP/3 SETTINGS in wire format, so that early
data can be accepted on a connection only if they match the ones from
the original connection.

However this is slightly annoying as it would require re-encoding
transport parameters _excluding_ the ones that shouldn't be remembered
when using 0-RTT, as well as exposing some kind of API to allow the
HTTP/3 layer to pass the SETTINGS blob, and in practice these parameters
don't really change that often, so leave that as a TODO item.