stream: only validate stream id of new streams if they never existed
When a STREAM frame is received and `get_or_create()` is called, we
assume that the stream was not created locally.

However it can happen that a locally-created stream was already
collected and then duplicated data for it is received.

This means that we should check whether the stream was destroyed, before
validating the stream ID.