qpack: avoid multiple copies when decoding literals
When literal names and values are decoded, they are additionally cloned
when Header::new() is called. However in many cases these values are
already of type String, which means that the additional clone is
unnecessary, so try to create Header objects directly when possible to
avoid such clones.

In the case of literal names, the source buffer was additionally cloned
in case of plain strings (i.e. not huffman-encoded), so avoid that as
well.