qns: specify wire version in the client
Currently we always trigger a version negotiation, but that's not going
to work anymore due to https://github.com/marten-seemann/quic-interop-runner/pull/229.

Instead we specify an explicit wire version, which should match the
version expected by the interop runner.

An alternative would be to make the client always select the latest
supported version by default, which would avoid having to update this
whenever a new draft is implemented.

But on the other hand we would still need to match what the runner
supports, so this will allow us to potentially add new draft versions
without having to wait for the runner to support it. Also, this is a lot
easier to implement :)